# contracts/mfdataindia-openapi-v2.json
{
  "components": {
    "schemas": {
...
      "CompareResponse": {
        "description": "GET /api/compare \u2014 identities are retained, while consumers use the\nsupplied policy objects to withhold ineligible normalized performance.",
        "properties": {
          "funds": {
            "items": {
              "$ref": "#/components/schemas/ComparisonItem"
            },
            "title": "Funds",
            "type": "array"
          },
          "years": {
            "title": "Years",
            "type": "number"
          }
        },
        "required": [
          "years",
          "funds"
        ],
        "title": "CompareResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PeersResponse": {
        "description": "GET /api/funds/{code}/peers \u2014 peer population excludes\ncomparison-ineligible series.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "horizons": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PeerHorizon"
            },
            "title": "Horizons",
            "type": "object"
          }
        },
        "required": [
          "code",
          "horizons"
        ],
        "title": "PeersResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MFDataIndia",
    "version": "0.1.0",
    "x-contract-id": "mfdataindia-json-read-v2",
    "x-contract-version": 2
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/compare": {
      "get": {
        "operationId": "api_compare_api_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "codes",
            "required": true,
            "schema": {
              "maxLength": 128,
              "title": "Codes",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "years",
            "required": false,
            "schema": {
              "default": 1.0,
              "maximum": 100,
              "minimum": 0,
              "title": "Years",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Compare"
      }
    },
    "/api/fund-families": {
      "get": {
        "operationId": "api_fund_families_api_fund_families_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "amc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Amc"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "option",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Option"
            }
          },
          {
            "in": "query",
            "name": "live",
            "required": false,
            "schema": {
              "default": true,
              "title": "Live",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name",
              "maxLength": 32,
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plan",
            "required": false,
            "schema": {
              "default": "regular",
              "maxLength": 16,
              "title": "Plan",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Fund Families Api Fund Families Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Fund Families"
      }
    },
    "/api/funds": {
      "get": {
        "operationId": "api_funds_api_funds_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "amc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Amc"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "option",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Option"
            }
          },
          {
            "in": "query",
            "name": "in_scope",
            "required": false,
            "schema": {
              "default": true,
              "title": "In Scope",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "live",
            "required": false,
            "schema": {
              "default": true,
              "title": "Live",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name",
              "maxLength": 32,
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plan",
            "required": false,
            "schema": {
              "default": "regular",
              "maxLength": 16,
              "title": "Plan",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Funds Api Funds Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Funds"
      }
    },
    "/api/funds/batch": {
      "get": {
        "operationId": "api_funds_batch_api_funds_batch_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Funds Batch Api Funds Batch Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Funds Batch"
      }
    },
    "/api/funds/{code}": {
      "get": {
        "operationId": "api_fund_api_funds__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "plan",
            "required": false,
            "schema": {
              "default": "regular",
              "maxLength": 16,
              "title": "Plan",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Fund"
      }
    },
    "/api/funds/{code}/analytics": {
      "get": {
        "operationId": "api_fund_analytics_api_funds__code__analytics_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Fund Analytics"
      }
    },
    "/api/funds/{code}/nav": {
      "get": {
        "operationId": "api_fund_nav_api_funds__code__nav_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "years",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Years"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Fund Nav Api Funds  Code  Nav Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Fund Nav"
      }
    },
    "/api/funds/{code}/peers": {
      "get": {
        "operationId": "api_fund_peers_api_funds__code__peers_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Fund Peers"
      }
    },
    "/api/funds/{code}/returns": {
      "get": {
        "operationId": "api_fund_returns_api_funds__code__returns_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Fund Returns"
      }
    },
    "/api/funds/{code}/risk-reward": {
      "get": {
        "operationId": "api_fund_risk_reward_api_funds__code__risk_reward_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskRewardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Fund Risk Reward"
      }
    },
...
  },
...
}